// frontend/src/types/api.d.ts
export interface ProblemDetailError {
  field: string;
  code: string;
  message: string;
}

export interface ProblemDetail {
  type: string;
  title: string;
  status: number;
  code: string;
  errors?: ProblemDetailError[];
  traceId?: string;
}